refactor: 파일 구조 변경 MVC 방향으로 변경
