+++ src/js/core/reactive/index.js
