+++ src/js/view/LottoItem.js
